api_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
moviepy==1.0.3
faster-whisper
python-dotenv
opencv-python-headless
numpy
pillow
requests
arabic-reshaper
python-bidi
imageio==2.25.1
imageio-ffmpeg==0.4.8
openai>=1.0.0
scipy
json_repair
tiktoken
pydantic
deep-translator